requirements.txt
absl-py==2.1.0
albucore==0.0.16
albumentations==1.4.18
annotated-types==0.7.0
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.0
cmake==3.31.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
eval_type_backport==0.2.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
grpcio==1.68.1
idna==3.10
imagecodecs==2024.9.22
imageio==2.36.0
ipython==8.30.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
lit==18.1.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.1
numcodecs==0.13.1
numexpr==2.10.1
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.4.0.58
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.6.0.99
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
Pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.29.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.0
torchvision==0.20.0
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13